requirements.txt
flask
opencv-python-headless
numpy
pillow
requests
python-dotenv
